refactor: 라우터 설정을 별도 파일로 분리
- src/app/router.tsx 파일로 라우터 생성 코드 추출
- main.tsx에서 라우터 관련 코드 제거 후 import로 대체
